airdrop_handlers.py
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import CallbackContext

# Assuming you have a function to get user-specific data
def get_user_points(user_id):
    # Implement the logic to fetch user's points
    return 100  # Example value

def get_user_volume(user_id):
    # Implement the logic to fetch user's volume
    return 1000  # Example value

def get_total_volume():
    # Implement the logic to fetch total volume
    return 10000  # Example value

async def handle_airdrop(update: Update, context: CallbackContext):
    query = update.callback_query
    if query:
        await query.answer()

    user_id = update.effective_user.id
    points_accumulated = get_user_points(user_id)
    user_volume = get_user_volume(user_id)
    total_volume = get_total_volume()

    message = (
        f"Points Accumulated: {points_accumulated}\n"
        f"Your Volume: {user_volume}\n"
        f"Total Volume: {total_volume}\n\n"
        "Note: Points will be convertible to $ZAP at TGE.\n\n"
        "1 Point = $1 In Volume"
    )

    reply_markup = InlineKeyboardMarkup([[InlineKeyboardButton("Back", callback_data='back_to_start')]])
    await query.edit_message_text(message, reply_markup=reply_markup)

